+++ src/main/java/com/MGR/dto/MemberFormDto.java
@@ -19,7 +19,8 @@ public class MemberFormDto {
     @Length(min=8, max=16, message="비밀번호는 8자 이상, 16자 이하로 입력해주세요")
     private String password;
 
-    @NotBlank(message = "비밀번호 확인을 입력하세요")
+    @NotEmpty(message = "비밀번호 확인은 필수 입력 값입니다")
+    @Length(min=8, max=16)
     private String password2;
 
     @NotEmpty(message = "별명은 필수 입력 값입니다")
