+++ src/main/resources/templates/member/joinForm.html
@@ -8,11 +8,6 @@
 
 <th:block layout:fragment="css">
     <style>
-
-    .email_ok{
-        color: blue;
-        display: none;
-    }
     .email_already{
         color: red;
         display: none;
@@ -25,11 +20,19 @@
         color: red;
         display: none;
     }
+    .nickname_null{
+        color: red;
+        display: none;
+    }
     .password_ok{
         color: blue;
         display: none;
     }
     .password_already{
+        color: red;
+        display: none;
+    }
+    .password_null{
         color: red;
         display: none;
     }
@@ -49,20 +52,25 @@
         function checkEmail(){
             var email = $('#email').val(); // #email 에서 입력받은 값 변수 저장
             prevEmail = email;
+
+            $('.email_already').css("display", "none");
+            $('.email_error').css("display", "none");
+
             $.ajax({
                 url:'/member/emailCheck', //controller 에서 요청 받을 주소
                 type:'POST',
                 data:{email:email},
+                beforeSend : function(xhr){
+                    xhr.setRequestHeader(header, token);
+                },
                 success:function(result){
-
                 //controller 에서 넘어온 result 값이 1이면 중복
+                console.log("result= " + result)
                     if(result == 0){
-                        $('.email_ok').css("display","inline-block");
                         $('.email_already').css("display", "none");
                         $('.email_error').css("display", "none");
                     } else {
                         $('.email_already').css("display","inline-block");
-                        $('.email_ok').css("display", "none");
                         $('#email').val(prevEmail);
                     }
 
@@ -72,7 +80,7 @@
 
         var prevNickname = "";
         function checkNickname(){
-            var nickname = $('#nickname').val();
+            var nickname = $('#nickname').val().trim();
             prevNickname = nickname;
             $.ajax({
                 url:'/member/nicknameCheck',
@@ -81,8 +89,7 @@
                 beforeSend : function(xhr){
                     xhr.setRequestHeader(header, token);
                 },
-                success:function(){
-
+                success:function(result){
                     if(result === 0){
                         $('.nickname_ok').css("display","inline-block");
                         $('.nickname_already').css("display", "none");
@@ -91,7 +98,13 @@
                         $('.nickname_ok').css("display", "none");
                         $('#nickname').val(prevNickname);
                     }
-
+                    if(!nickname){
+                        $('.nickname_null').css("display","inline-block");
+                        $('.nickname_ok').css("display", "none");
+                        $('.nickname_already').css("display", "none");
+                    }else{
+                        $('.nickname_null').css("display","none");
+                    }
                 }
             });
         };
@@ -118,6 +131,13 @@
                         $('.password_already').css("display", "inline-block");
                         $('.password_ok').css("display", "none");
                     }
+                    if(!password){
+                        $('.password_null').css("display","inline-block");
+                        $('.password_ok').css("display", "none");
+                        $('.password_already').css("display", "none");
+                    }else{
+                        $('.password_null').css("display","none");
+                    }
                 }
             });
         }
@@ -140,7 +160,6 @@
             });
         }
 
-
     </script>
 </th:block>
 
@@ -149,18 +168,17 @@
     <form action="/member/create" role="form" th:object="${memberFormDto}" method="POST">
         <div class="mb-3">
             <label th:for="email">이메일</label>
-            <input th:field="*{email}" class="form-control" type="email" onblur="checkEmail()"/>
-<!--            <span class="email_ok">사용 가능한 이메일 입니다</span>-->
-            <span class="email_already">이미 사용증인 이메일 입니다</span>
+            <input th:field="*{email}" id="email" class="form-control" type="email" onblur="checkEmail()"/>
+            <span class="email_already">이미 사용중인 이메일 입니다</span>
             <span class="validationError" th:if="${#fields.hasErrors('email')}"
                   th:errors="*{email}"></span>
-            <p class="emailError" th:text="${errors}"></p>
         </div>
         <div class="mb-3">
             <label for="password">비밀번호</label>
-            <input class="form-control" type="password" id="password" th:field="*{password}"/>
+            <input class="form-control" type="password" id="password" th:field="*{password}" onblur="checkPassword()"/>
             <span class="validationError" th:if="${#fields.hasErrors('password')}"
                   th:errors="*{password}"></span>
+            <span class="nickname_null">비밀번호를 입력하세요</span>
         </div>
         <div class="mb-3">
             <label for="password2">비밀번호 확인</label>
@@ -181,6 +199,7 @@
             <input class="form-control" type="text" id="nickname" th:field="*{nickname}" onkeyup="checkNickname()"/>
             <span class="nickname_ok">사용 가능한 닉네임 입니다</span>
             <span class="nickname_already">이미 사용중인 닉네임 입니다</span>
+            <span class="nickname_null">닉네임을 입력하세요</span>
             <span class="validationError" th:if="${#fields.hasErrors('nickname')}"
                   th:errors="*{nickname}"></span>
         </div>
@@ -192,8 +211,10 @@
         </div>
 
         <div class="mb-3">
-            <label for="nickname">메일인증</label>
+            <label for="code">메일인증 코드</label>
             <input class="form-control" type="text" id="code" th:field="*{code}"/>
+            <span class="validationError" th:if="${#fields.hasErrors('code')}"
+                  th:errors="*{code}"></span>
             <button class="btn btn-primary" type="button" id="code-button" onclick="sendEmail()">인증 메일 발송</button>
             <input type="hidden" name="authCode" th:value="*{authCode}">
         </div>
@@ -201,7 +222,6 @@
         <button class="btn btn-primary" type="submit" id="member-create">가입</button>
         <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
     </form>
-
 </div>
 </html>
