+++ src/main/resources/templates/member/loginForm.html
@@ -16,7 +16,7 @@
 
     <div class="container my-3">
         <h5 class="my-3 border-bottom pb-2">로그인</h5>
-        <form th:action="@{/member/login}" method="POST">
+        <form th:action="@{/login}" method="POST">
             <div class="mb-3">
                 <label for="email"> 이메일</label>
                 <input class="form-control" type="text" name="email" id="email"/>
